v0.0.0.8 - Creating more structures
This time we started to create the structure of modules, grids and themes that will be part of our project's system. Modules are code that can be embedded into "plug-in"-like code. The "grids" will be the pages that the user will create through the dashboard. "Themes" are the layouts the system user will choose for their web page