v0.0.2.3 - Using classes
This time, we start embedding the classes we created in our "class" folder within the system. We also started to fix small bugs that had been left in previous versions of the code, we fixed that problem. We also took the first step to build the user panel elements from the admin.php file.