v0.0.3.7 - Creating Page Controls
This time we create a file to control the execution and display of pages inside the gridpage administration panel, the panel-page.php file will be in charge of managing this control. We also made some punctual changes inside the code adding some javascript functions that will be in charge of showing and hiding dynamic elements inside the administration panel, we made some small punctual changes to make the function names more integrated with the system as a whole.