v0.0.3.8 - Adding features
This time we started to implement some new features where we started to create a new file inside the system version directory, the file name is "controller.php", this file will be in charge of controlling the execution flow of the application as a whole making calls to classes and other elements that need to be controlled during site and system execution. We also made minor changes in some points in the system and fixed a small error in the declaration of one of the functions.